fix: refactor commit log
