+++ packages/at_persistence_secondary_server/test/hive_key_expiry_check.dart
@@ -1,6 +1,8 @@
-import 'dart:io';
-
+import 'package:at_commons/at_commons.dart';
 import 'package:at_persistence_secondary_server/at_persistence_secondary_server.dart';
+import 'package:at_persistence_secondary_server/src/log/commitlog/server/at_server_commit_log.dart';
+import 'package:at_persistence_secondary_server/src/log/commitlog/server/at_server_commit_log_keystore.dart';
+import 'package:test/expect.dart';
 
 void main() async {
   var secondaryPersistenceStore = SecondaryPersistenceStoreFactory.getInstance()
@@ -12,27 +14,22 @@ void main() async {
   var keyStoreManager =
       secondaryPersistenceStore.getSecondaryKeyStoreManager()!;
   var keyStore = secondaryPersistenceStore.getSecondaryKeyStore()!;
-  var commitLogKeyStore = CommitLogKeyStore('@test_user_1');
+  var commitLogKeyStore = AtServerCommitLogKeyStore('@test_user_1');
   await commitLogKeyStore.init('test/hive/commit');
-  keyStore.commitLog = AtCommitLog(commitLogKeyStore);
+  keyStore.commitLog = AtServerCommitLog(commitLogKeyStore);
   keyStoreManager.keyStore = keyStore;
   var atData = AtData();
   atData.data = 'abc';
   await keyStoreManager
       .getKeyStore()
-      .put('123', atData, time_to_live: 30 * 1000);
-  print('end');
-  var atDataResponse = await keyStoreManager.getKeyStore().get('123');
+      .put('phone.wavi@test_user_1', atData, time_to_live: 2000);
+  var atDataResponse =
+      await keyStoreManager.getKeyStore().get('phone.wavi@test_user_1');
   print(atDataResponse?.data);
   assert(atDataResponse?.data == 'abc');
-  var expiredKey =
-      await Future.delayed(Duration(minutes: 2), () => getKey(keyStoreManager));
-  assert(expiredKey == null);
-  print(expiredKey);
-  exit(0);
-}
-
-Future<String?> getKey(keyStoreManager) async {
-  AtData? atData = await keyStoreManager.getKeyStore().get('123');
-  return atData?.data;
+  await Future.delayed(Duration(seconds: 60));
+  expect(
+      () async =>
+          await keyStoreManager.getKeyStore().get('phone.wavi@test_user_1'),
+      throwsA(predicate((dynamic e) => e is KeyNotFoundException)));
 }
